Move the feature list off the front page.
Generally start referring to features rather than feature files.
Decouple features storage directory from Express route URLs.
Move the feature list off the front page.
Move the feature related routes into a single Express routing module.
Make the feature tests more rigorous (won't pass on an error message with 'feature' in it).
Make the css paths on the feature list page relative to the server route rather than current page (fixes bug on trailing slash in features/ URL).
