+++ package.json
@@ -6,7 +6,7 @@
     "prestart": "npm stop",
     "start": "node ./bin/www",
     "stop": "pkill specs || true",
-    "pretest": "npm start > /dev/null & sleep 1",
+    "pretest": "npm start & sleep 1",
     "test": "node_modules/.bin/cucumber-js --require features-support",
     "posttest": "npm stop",
     "pretest-with-coverage": "npm run pretest",
