+++ views/features.hbs
@@ -8,9 +8,9 @@
     <meta name="description" content="A service for exposing Gherkin feature files as specifications.">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <link rel="stylesheet" href="public/css/modified-normalize.css">
-    <link rel="stylesheet" href="public/css/alphabeta.css">
-    <link rel="stylesheet" href="public/css/main.css">
+    <link rel="stylesheet" href="/public/css/modified-normalize.css">
+    <link rel="stylesheet" href="/public/css/alphabeta.css">
+    <link rel="stylesheet" href="/public/css/main.css">
 
     <!-- Old IE hack for main elements-->
     <script>document.createElement('main');</script>
